+++ src/app/api/__test__/user.test.ts
@@ -2,7 +2,7 @@
 import { expect, test } from "bun:test";
 //@ts-ignore
 import { describe } from "bun:test";
-import { CartRequestImpl, FirebaseRequestImpl, ManufacturerRequestImpl, StripeRequestImpl, UserRequestImpl } from "../lib/firebase";
+import { CartRequestImpl, FirebaseRequestImpl, ManufacturerRequestImpl, StripeRequestImpl, UserRequestImpl } from "../lib/instances";
 import { auth } from "../firebase";
 import { Address } from "../models/user";
 import { Cart, CartRegisterPayload } from "../models/cart";
