+++ src/app/cart/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { useAuthContext } from '../contexts/AuthContext'
-import { CartRequestImpl } from '../api/lib/firebase'
+import { CartRequestImpl } from '../api/lib/instances'
 
 const Cart = () => {
   const user = useAuthContext()
