+++ src/app/stripe/page.tsx
@@ -4,7 +4,7 @@ import { loadStripe } from '@stripe/stripe-js'
 import { Elements } from '@stripe/react-stripe-js'
 
 import CheckoutForm from './_component/CheckoutForm'
-import { StripeRequestImpl } from '../api/lib/firebase'
+import { StripeRequestImpl } from '../api/lib/instances'
 import { useAuthContext } from '../contexts/AuthContext'
 import { IStripeRequests } from '../api/models/stripe'
